feat: implement --force support across all methods and modularize object copying
