+++ .omg/state/learn-watch.json
@@ -1,12 +1,12 @@
 {
   "last_session_id": "a1e53197-8618-4d65-ad59-07833a7fc4b9",
-  "last_event_key": "a1e53197-8618-4d65-ad59-07833a7fc4b9:transcript:9a42b4ad4fec4ff9f6e590aae8b8e3637b1c809dcef27cdc452c20c4629741d0",
+  "last_event_key": "a1e53197-8618-4d65-ad59-07833a7fc4b9:transcript:a9208bbd48c46c89ea10f1f92d11cf124ecc091e67ccdb7072eb36b866a881d5",
   "last_prompted_session_id": "a1e53197-8618-4d65-ad59-07833a7fc4b9",
   "last_reason": "prompt-cooldown-active",
   "last_prompted_at": "2026-04-23T03:46:51.329Z",
-  "last_user_message_count": 13,
+  "last_user_message_count": 14,
   "last_actionable_message_count": 4,
   "deep_interview_lock_active": false,
   "deep_interview_lock_source": "/home/dbado/sfDBTools/.omg/state/deep-interview.json",
-  "updated_at": "2026-04-23T04:15:02.159Z"
+  "updated_at": "2026-04-23T04:16:32.111Z"
 }
