+++ .omg/state/quota-watch.json
@@ -1,20 +1,20 @@
 {
-  "turn_count": 12,
+  "turn_count": 13,
   "last_session_id": "a1e53197-8618-4d65-ad59-07833a7fc4b9",
-  "last_event_key": "a1e53197-8618-4d65-ad59-07833a7fc4b9:transcript:9a42b4ad4fec4ff9f6e590aae8b8e3637b1c809dcef27cdc452c20c4629741d0",
-  "last_transcript_hash": "9a42b4ad4fec4ff9f6e590aae8b8e3637b1c809dcef27cdc452c20c4629741d0",
+  "last_event_key": "a1e53197-8618-4d65-ad59-07833a7fc4b9:transcript:a9208bbd48c46c89ea10f1f92d11cf124ecc091e67ccdb7072eb36b866a881d5",
+  "last_transcript_hash": "a9208bbd48c46c89ea10f1f92d11cf124ecc091e67ccdb7072eb36b866a881d5",
   "last_model": "gemini-3-flash-preview",
   "last_provider": "gemini",
-  "last_turn_total_tokens": 404773,
-  "last_session_total_tokens": 28065213,
+  "last_turn_total_tokens": 407134,
+  "last_session_total_tokens": 29690806,
   "session_totals_by_model": {
     "gemini-3.1-pro-preview": 1730068,
-    "gemini-3-flash-preview": 26335145
+    "gemini-3-flash-preview": 27960738
   },
   "session_totals_by_provider": {
-    "gemini": 28065213
+    "gemini": 29690806
   },
   "cwd_mode": "parent-leaf",
   "last_cwd_label": "dbado/sfDBTools",
-  "updated_at": "2026-04-23T04:15:02.180Z"
+  "updated_at": "2026-04-23T04:16:32.140Z"
 }
